test: Move MockHTTPTextFormat to test package
This mock propagator could be use for testing propagation in different packages.
It was implemented in the opentracing-shim and this commit moves it to a generic
place.
