+++ ext/opentelemetry-ext-opentracing-shim/setup.cfg
@@ -44,5 +44,9 @@ install_requires =
     opentracing ~= 2.0
     opentelemetry-api == 0.7.dev0
 
+[options.extras_require]
+test =
+    opentelemetry-ext-testutil == 0.7.dev0
+
 [options.packages.find]
 where = src
