+++ ext/opentelemetry-ext-opentracing-shim/tests/test_shim.py
@@ -16,26 +16,15 @@
 # pylint:disable=no-member
 
 import time
-import typing
 from unittest import TestCase
 
 import opentracing
 
 import opentelemetry.ext.opentracing_shim as opentracingshim
 from opentelemetry import propagators, trace
-from opentelemetry.context import Context
 from opentelemetry.ext.opentracing_shim import util
+from opentelemetry.ext.testutil.mock_httptextformat import MockHTTPTextFormat
 from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.trace.propagation import (
-    get_span_from_context,
-    set_span_in_context,
-)
-from opentelemetry.trace.propagation.httptextformat import (
-    Getter,
-    HTTPTextFormat,
-    HTTPTextFormatT,
-    Setter,
-)
 
 
 class TestShim(TestCase):
@@ -542,46 +531,3 @@ class TestShim(TestCase):
         # Verify exception for non supported binary format.
         with self.assertRaises(opentracing.UnsupportedFormatException):
             self.shim.extract(opentracing.Format.BINARY, bytearray())
-
-
-class MockHTTPTextFormat(HTTPTextFormat):
-    """Mock propagator for testing purposes."""
-
-    TRACE_ID_KEY = "mock-traceid"
-    SPAN_ID_KEY = "mock-spanid"
-
-    def extract(
-        self,
-        get_from_carrier: Getter[HTTPTextFormatT],
-        carrier: HTTPTextFormatT,
-        context: typing.Optional[Context] = None,
-    ) -> Context:
-        trace_id_list = get_from_carrier(carrier, self.TRACE_ID_KEY)
-        span_id_list = get_from_carrier(carrier, self.SPAN_ID_KEY)
-
-        if not trace_id_list or not span_id_list:
-            return set_span_in_context(trace.INVALID_SPAN)
-
-        return set_span_in_context(
-            trace.DefaultSpan(
-                trace.SpanContext(
-                    trace_id=int(trace_id_list[0]),
-                    span_id=int(span_id_list[0]),
-                    is_remote=True,
-                )
-            )
-        )
-
-    def inject(
-        self,
-        set_in_carrier: Setter[HTTPTextFormatT],
-        carrier: HTTPTextFormatT,
-        context: typing.Optional[Context] = None,
-    ) -> None:
-        span = get_span_from_context(context)
-        set_in_carrier(
-            carrier, self.TRACE_ID_KEY, str(span.get_context().trace_id)
-        )
-        set_in_carrier(
-            carrier, self.SPAN_ID_KEY, str(span.get_context().span_id)
-        )
